requirements.txt
fastapi
pydantic
google-genai
